Do not connect to file pk
What happens here is that it uses the id from grouper and it tries
to use that id to join with the file table id.

The problem is that grouper_id is not the same as the file pk so
you'll end up with a file you weren't looking for.

So instead we join on grouper id that exists on the file table.